% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateMod.R
\name{CreateMod}
\alias{CreateMod}
\title{Create Adjunct CmdStan Model Environment}
\usage{
CreateMod(
  coef_hyper,
  sd_hyper,
  nWarmup_init,
  nSamples_init,
  nWarmup_run,
  nSamples_run,
  aux_envir
)
}
\arguments{
\item{coef_hyper}{Hyperparameter value for the standard deviation of normally distributed parameters}

\item{sd_hyper}{Hyperparameter value for the shape parameter of gamma distributed parameters}

\item{nWarmup_init}{Number of burn-in draws for the fixed \mjeqn{\theta}{} model}

\item{nWarmup_run}{Number of burn-in draws for the free \mjeqn{\theta}{} model}

\item{nSamples_run}{Number of sampled posterior values for the free \mjeqn{\theta}{} model (after burn-in has completed)}

\item{aux_envir}{An environment object that contains objects to be loaded into the \code{CreateMod} environment (such as is returned by the \code{\link{genData}} function)}

\item{nSampels_init}{Number of posterior draws for the fixed \mjeqn{\theta}{} model (after burn-in has completed)}
}
\value{
an environment object containing fixed \mjeqn{\theta}{} and free \mjeqn{\theta}{} model results
}
\description{
Creates an environment to estimate the \mjeqn{p \times i}{} interaction model with CmdStan.
}
\details{
\loadmathjax{}

Creates an environment with methods for fitting the fixed \mjeqn{\theta}{} and free \mjeqn{\theta}{} models. Additionally, \code{CreateMod} includes a method for recursively checking \mjeqn{\hat{R}}{} convergence.
}
\seealso{
\code{\link{initialize}}, \code{\link{sample}}, \code{\link{rhatCheck}}
}
