% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkcmdstan.R
\name{.onLoad}
\alias{.onLoad}
\title{Load/Install the cmdstanr package}
\usage{
.onLoad(...)
}
\description{
Validates installation of the cmdstanr package and, if not found, prompts for its installation from the recommended source
}
\details{
\loadmathjax{}

See https://mc-stan.org/cmdstanr/
}
