% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genData.R
\name{genData}
\alias{genData}
\title{Data Generating Function for the \mjeqn{p \times i}{} Interaction Model}
\usage{
genData(P, I, J, K, seed = NULL, isCorrI = TRUE)
}
\arguments{
\item{P}{Number of examinees}

\item{I}{Number of items}

\item{J}{Number of potential item feature categories}

\item{K}{Number of potential person feature categories}

\item{seed}{Integer seed for replication (if \code{NULL}, a random seed will be generated)}
}
\value{
an environment containing simulated data
}
\description{
Generates data for a \mjeqn{p \times i}{} Interaction Model under the Explanatory Item Response theoretical framework.
}
\details{
\loadmathjax{}

Let a model of the log-odds transformed outcome \mjeqn{(\eta_{pi})}{} be defined as:
\mjdeqn{\eta_{pi} = \theta_{p}\lambda_{i} + \tau_{i} + \sum_{k=1}^{K}\sum_{j=1}^{J}\beta_{jk_{(\eta)}}x_{ik}z_{pj}}{ representation}{}
where the linear predictor of latent trait measurements \mjeqn{\theta_{p}}{}is defined as:
\mjdeqn{\theta_{p} = \sum_{j=1}^{J}\beta_{j_{(\theta)}}z_{pj} + u_{p_{(\theta)}}}{}
where the linear predictor of item slope/discrimination terms (\mjeqn{\lambda_{i}}{}) is defined as:
\mjdeqn{\lambda_{i} = \sum_{k=1}^{K}\beta_{k_{(\lambda)}}x_{ik} + u_{i_{(\lambda)}}}{}
and where the linear predictor of item intercept/easiness terms (\mjeqn{\tau_{i}}{}) is defined as:
\mjdeqn{\tau_{i} = \sum_{k=1}^{K}\beta_{k_{(\tau)}}x_{ik} + u_{i_{(\tau)}}}{}
For the linear predictor of latent trait measurements, the following assumption is made on its residual terms for the purposes of model identification:
\mjdeqn{u_{p_{(\theta)}} \sim \mathcal{N}(0,1)}{}
If an assumption is made that there is no strcutured dependency between the residual terms for the linear predictors of item intercepts and item slopes (\code{isCorrI = FALSE}), then:
\mjdeqn{\begin{gathered} u_{i_{(\lambda)}} \sim \mathcal{N}(0, \sigma^{2}_{\lambda}) \cr u_{i_{(\tau)}} \sim \mathcal{N}(0, \sigma^{2}_{\tau})\end{gathered}}{}
Whereas if it is assumed that the residuals of the linear predictors of item intercepts and item slopes is assumed to have a predictable underlying structure (\code{isCorrI = TRUE}), then:
\mjdeqn{\begin{bmatrix}
u_{i_{(\lambda)}} \cr u_{i_{(\tau)}}
\end{bmatrix}
\sim \mathcal{MVN}
\Bigg(\mu =
       \begin{bmatrix} 0 \cr 0 \end{bmatrix}
     , \hspace{2pt} \mathbf{\Sigma} =
       \begin{bmatrix}\sigma^{2}_{\lambda} & \sigma_{\lambda}\sigma_{\tau} \cr \sigma_{\tau}\sigma_{\lambda} & \sigma^{2}_{\tau}
     \end{bmatrix}
     \Bigg)}{}
}
