% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genReport.R
\name{genReport}
\alias{genReport}
\title{Generate Text Report of Free \mjeqn{\theta}{} Model Results}
\usage{
genReport(saveDir, fileDetails, ...)
}
\arguments{
\item{saveDir}{Path to directory where report is saved}

\item{fileDetails}{string identifier for name of saved report}

\item{...}{additional arguments passed from parent frame}
}
\value{
a \code{.txt} file of the free \mjeqn{\theta}{} results
}
\description{
Helper function that generates a text document in the user-specified save directory
}
\details{
\loadmathjax{}

Model results include posterior descriptives and simulated data information
}
