% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Run Fixed \mjeqn{\theta}{} Model}
\usage{
initialize(...)
}
\arguments{
\item{...}{arguments passed from parent frame}
}
\value{
Model results for the standardized \mjeqn{\theta}{} model added to the environment rendered by \code{\link{CreateMod}}
}
\description{
Runs the first step in the two-step estimation framework
}
\details{
\loadmathjax{}

The first step in the two-step estimation framework for the \mjeqn{p \times i}{} model
}
\seealso{
\code{\link{CreateMod}}
}
