% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhatCheck.R
\name{rhatCheck}
\alias{rhatCheck}
\title{Recursive Validation for Rhat Convergence}
\usage{
rhatCheck(...)
}
\arguments{
\item{...}{arguments passed from parent frame}
}
\value{
modified free \mjeqn{\theta}{} model that is added to an environment created by \code{\link{CreateMod}}
}
\description{
Helper function that validates \mjeqn{\hat{R}}{} convergence.
}
\details{
\loadmathjax{}

\deqn{
\begin{array}{l}
\textbf{Input: } \texttt{CmdStan}\hspace{3pt}\text{model,}\hspace{3pt}N_{\mathrm{max}}\text{,}\hspace{3pt}\hat{R}_{1} \\
\text{Sample }\texttt{CmdStan}\hspace{3pt}\text{model} \\
\quad \textbf{while}\hspace{5pt} \hat{R}_{n}\leq\hat{R}_{thresh}\hspace{5pt} \textbf{or} \hspace{5pt} n \leq N_{\mathrm{max}}\hspace{5pt}\textbf{do} \\
\qquad \text{Sample }\texttt{CmdStan}\hspace{3pt}\text{model} \\
\qquad n \mathrel{+}= 1 \\
\quad \textbf{end do} \\
\textbf{end while} \\
\end{array}
}
}
\references{
Vehtari, A., Gelman, A., Simpson, D., Carpenter, B., & Bürkner, P. (2021). Rank-normalization, folding, and localization: an improved rˆ for assessing convergence of mcmc (with discussion). Bayesian Analysis, 16(2). https://doi.org/10.1214/20-ba1221
}
\seealso{
\code{\link{CreateMod}}
}
