% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample}
\alias{sample}
\title{Run Free \mjeqn{\theta}{} Model}
\usage{
sample(...)
}
\arguments{
\item{...}{arguments passed from parent frame}
}
\value{
Model results for the full model added to the environment rendered by \code{\link{CreateMod}}
}
\description{
Runs the second step of the two-step estiamtion framework
}
\details{
\loadmathjax{}

The second step treating latent trait measurements (\mjeqn{\theta_{p}}{}) as freely estimated parameters in the two-step estimation framework for the \mjeqn{p \times i}{} model
}
\seealso{
\code{\link{CreateMod}}
}
