% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCorr.R
\name{rCorr}
\alias{rCorr}
\title{Randomly Sample a Cholesky Factorized, Lower Triangular Matrix from a Lewandowski-Kurowicka-Joe Distribution}
\usage{
rCorr(nDim, eta = 1)
}
\arguments{
\item{nDim}{desired dimension of the sampled \mjeqn{n \times n}{} matrix}

\item{eta}{concentration hyperparameter}
}
\value{
a lower triangular \mjeqn{n \times n}{} matrix
}
\description{
Generates a lower triangular, invertible, positive-definite matrix
}
\details{
\loadmathjax{}

Using the onion method, samples a Cholesky factorized lower-triangular matrix \mjeqn{L}{} from a Lewandowski-Kurowicka-Joe (LKJ) distribution given the concentration hyperparameter \mjeqn{\eta}{}, such that: \mjdeqn{A=LL^{\top}}{} where \mjeqn{A}{} is a \mjeqn{n \times n}{} matrix and is a member of the set of all symmetric, positive-definite matrices.
}
