% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rCorr.R
\name{rCorr}
\alias{rCorr}
\title{Generation of Data for the Person-by-Item Interaction Model}
\usage{
rCorr(nDim, eta = 1)
}
\arguments{
\item{nDim}{desired dimension of the sampled \mjeqn{n \times n}{} matrix}

\item{eta}{concentration hyperparameter}
}
\value{
a Choleksy-factorized, lower triangular matrix
}
\description{
randomly samples a lower triangular, Cholesky-factorized matrix \mjeqn{L}{} from a Lewandowski-Kurowicka-Joe (LKJ) distribution given \mjeqn{\eta}{}.
}
\details{
\loadmathjax{}

Using the onion method, samples a Cholesky factorized lower-triangular matrix \mjeqn{L}{} from a Lewandowski-Kurowicka-Joe (LKJ) distribution given the concentration hyperparameter \mjeqn{\eta}{}, such that: \mjdeqn{A=LL^{\top}}{} where \mjeqn{A}{} is a \mjeqn{n \times n}{} matrix and is a member of the set of all symmetric, positive-definite matrices.
}
